backend/requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.32.0
python-dotenv==1.0.1
openai==1.42.0
supabase==2.4.5
pydantic==2.9.2
python-multipart==0.0.17
pillow==10.4.0
PyPDF2==3.0.1
docx2txt==0.8
scikit-learn==1.5.2
numpy==2.1.2
